chore: fix compiler warnings
This was quite noticable on the stdout/stderr that it made it hard to
notice the real error if there was any, in general it's a good idea to
explicitly mark parameters/variables unused as they are added in
empty/placeholder functions.

+++ projects/editor/source/layout/layout.c
@@ -9,11 +9,11 @@ status editor_count_rows(struct editor_state *editor) {
 	}
 
 	struct window *window = editor->window;
-	if (window->width < editor->font_size || window->height < editor->font_size) {
+	if (window->width < (i32) editor->font_size || window->height < (i32) editor->font_size) {
 		fprintf(stderr, "window size not large enough to render rows\n");
 		return status_failure;
 	}
 
-	editor->screen_rows = (u32) (window->height / editor->font_size);
+	editor->screen_rows = (u32) ((f32) window->height / editor->font_size);
 	return status_success;
 }
